feat: industrial-grade hardening, fractional PDC, and deep DNA integration
- Implement high-fidelity fractional Plugin Delay Compensation (PDC) using 4-point Lagrange interpolation across both sequential and parallel execution paths.
- Upgrade engine telemetry with high-resolution node-level resource accounting, including peak and average cycle tracking.
- Vectorize legacy Biquad filter unrolled paths for increased signal processing throughput.
- Implement DNA-based Groove Transfusion: automatically extract and apply micro-timing offsets from SoundDNA rhythmic profiles to sequencer patterns.
- Ensure strict real-time safety by moving all high-pressure allocations (PDC lines, crossfade buffers) to the Orchestration plane pre-allocation phase.
- Standardize all structural morphs and gain ramps to constant-power square-root curves.

+++ crates/nullherz-topology/src/compiler.rs
@@ -232,7 +232,7 @@ impl GraphCompiler {
                 if let Some(u) = v_to_producer[v_buf] {
                     let u_path_lat = path_latencies[u] + plan.node_latencies[u];
                     if max_v_path_lat > u_path_lat {
-                        plan.input_delays[v].0[i] = max_v_path_lat - u_path_lat;
+                        plan.input_delays[v].0[i] = (max_v_path_lat - u_path_lat) as f32;
                     }
                 }
             }
